fix change weight params to float

+++ main.py
@@ -34,35 +34,35 @@ async def save_imagery(file: UploadFile = File(...)):
 
 
 @app.get('/find/lang/html')
-async def find_by_language_return_html(content: str, relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_language_return_html(content: str, relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
     return HTMLResponse(html_response(thoughts))
 
 
 @app.get('/find/lang/text')
-async def find_by_language_return_plaintext(content: str, relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_language_return_plaintext(content: str, relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
     return PlainTextResponse(plaintext_response(thoughts))
 
 
 @app.get('/find/lang/file')
-async def find_by_language_return_file(content: str, relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_language_return_file(content: str, relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
     return FileResponse(file_response(thoughts))
 
 
 @app.get('/find/lang/json')
-async def find_by_language_return_json(content: str, relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_language_return_json(content: str, relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
     return json_response(thoughts)
 
 
 @app.post('/find/imag/html')
-async def find_by_imagery_return_html(file: UploadFile = File(...), relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_imagery_return_html(file: UploadFile = File(...), relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     content = await file.read()
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
@@ -70,7 +70,7 @@ async def find_by_imagery_return_html(file: UploadFile = File(...), relatedness:
 
 
 @app.post('/find/imag/text')
-async def find_by_imagery_return_plaintext(file: UploadFile = File(...), relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_imagery_return_plaintext(file: UploadFile = File(...), relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     content = await file.read()
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
@@ -78,7 +78,7 @@ async def find_by_imagery_return_plaintext(file: UploadFile = File(...), related
 
 
 @app.post('/find/imag/file')
-async def find_by_imagery_return_file(file: UploadFile = File(...), relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_imagery_return_file(file: UploadFile = File(...), relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     content = await file.read()
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
@@ -86,7 +86,7 @@ async def find_by_imagery_return_file(file: UploadFile = File(...), relatedness:
 
 
 @app.post('/find/imag/json')
-async def find_by_imagery_return_json(file: UploadFile = File(...), relatedness: Optional[int] = 1, activation: Optional[int] = 0, noise: Optional[int] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
+async def find_by_imagery_return_json(file: UploadFile = File(...), relatedness: Optional[float] = 1, activation: Optional[float] = 0, noise: Optional[float] = 0.05, silent: Optional[bool] = False, top_k: Optional[int] = 50):
     content = await file.read()
     thoughts = find(content, model, relatedness,
                     activation, noise, silent, top_k)
